Renamed compat to deprecated and removed files that are no longer needed
