--- glue/core/data_factories/io.py
@@ -1,6 +1,0 @@
-from __future__ import absolute_import, division, print_function
-
-# Backward-compatibility
-from .compat import extract_data_hdf5, extract_data_fits
-from .hdf5 import extract_hdf5_datasets
-from .fits import filter_hdulist_by_shape
